Migrate ClientScopeTest
Part of: #34494
