+++ tests/base/pom.xml
@@ -92,6 +92,12 @@
         </dependency>
         <dependency>
             <groupId>org.keycloak.tests</groupId>
+            <artifactId>keycloak-tests-utils-shared</artifactId>
+            <version>${project.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.keycloak.tests</groupId>
             <artifactId>keycloak-tests-custom-providers</artifactId>
             <version>${project.version}</version>
         </dependency>
