+++ tests/utils-shared/src/main/java/org/keycloak/testsuite/util/RoleBuilder.java
